Fixed a small bug.

+++ data/collect_demo_greedy_dagger.py
@@ -37,10 +37,10 @@ def greedy_state_selection(states, student_actions,
     action_error_trs = []
     action_error_rots = []
     for i, state in enumerate(states):
-        expert_action = expert_actions[i]
+        expert_action = expert_actions[i].astype(np.float64)
         expert_action[3:6] *= np.pi/180.
 
-        student_action = student_actions[i]
+        student_action = student_actions[i].astype(np.float64)
         student_action[3:6] *= np.pi/180.
 
         error = determine_action_error(student_action, expert_action, a1=a1, a2=a2)
